+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## v0.7.0
-- COD7-001: Provide Visit Hour Parts from Backend to enable filtering (#26) [86]
+- COD7-001: Provide Visit Hour Parts from Backend to enable filtering (#26) [105]
 
 ## v0.6.0 (2026-02-22)
 - COD6-001: Pick Map to fill coordinates on create location + delete btn (#17) [148]