+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - COD7-001: Provide Visit Hour Parts from Backend to enable filtering (#26) [105]
 - COD7-002: Add Search to Locations Page (#27) [64]
 - COD7-003: Add Collision Control on Travel Creation (#29) [101]
+- COD7-004: Allow bus line for bus type travels (#30) [101]
 
 ## v0.6.0 (2026-02-22)
 - COD6-001: Pick Map to fill coordinates on create location + delete btn (#17) [148]