+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - COD7-004: Allow bus line for bus type travels (#30) [101]
 - COD7-008: Migrate Zoom y Map Center to Statistics Backend (#28) [469]
 - COD7-012: Locations Responsive Page (#33) [664]
+- COD7-015: Show Unique Routes in Travel Footer (#36) [101]
 
 ## v0.6.0 (2026-02-22)
 - COD6-001: Pick Map to fill coordinates on create location + delete btn (#17) [148]